Pull the timeout poller to separate class
